### 2020-03-25 #23 @sabinomaggi
Reproduction of Step width enhancement in a pulse-driven Josephson junction
**Original article:** S. Maggi, _Step width enhancement in a pulse-driven Josephson junction_, Journal of Applied Physics vol. 79, no. 10, pp. 7860--7863 (1996), doi: [10.1063/1.362395](https://aip.scitation.org/doi/10.1063/1.362395).

**Scientific domain:**  Physics
**Programming language:**  Fortran/Visual Basic
