### 2020-03-31 #23 @pdebuyl
Hi @rougier

I can review but must ask first: 1. what is the current timeline? schedule is a mess (my schedule, of course) 2. I'd like to submit something as well, does that matter? (not the same topic of course)
