### 2020-04-13 #23 @rougier
@pdebuyl If you can review, that's all the better. I'll edit it then.
1. No real timeline but end of May would be nice
2. No problem for submitting something.
