### 2020-05-11 #23 @pdebuyl
Hi @sabinomaggi

The reproduction in the article is very interesting, thanks!

Before proceeding to the actual review, I would like to make the licensing checks. The procedure for the challenge (found [here](https://github.com/ReScience/ten-years/blob/master/author-guidelines.md) ) is a bit different from the usual ReScience one. There are still specific open-source concerns that are important.

1. Was there a license to the original code? Ideally, you should include the "old" code in the git commit history.
2. The new code must be licensed with an open license (see [debian's page](https://www.debian.org/social_contract#guidelines) for instance).
3. Do you have the source for the visual basic programs? All the software used in the replication must be made available.
