### 2020-05-11 #23 @sabinomaggi
Hi @pdebuyl
many thanks to you for the appreciation.

All current code has already been uploaded to this GitHub repository, https://github.com/sabinomaggi/ten-years-challenge-pulsed-drive. Now I see that I missed the last step, the submission to the Software Heritage archive. Is this enough to link everything together? Otherwise, how can I link the paper and the repository containing the code?

As for your specific questions:
1. If I understand well I should upload the "old" fortran code even if it is very similar to the current version. It is not a big problem, can I do that by using a branch?
2. The original code had no license, it was developed only to make the calculations I needed, not to be shared. In the GitHub repository mentioned above I have now licensed everything under CC4, I hope it is enough.
3. Also the original VB sources are already available in the GitHub repository.
