### 2020-05-13 #23 @pdebuyl
Apologies, I looked for a LICENSE file or a notice in the source files and did not see the notice in the README. That being said, creative commons is not suited for software, would you consider one of GPL, LGPL, BSD, MIT (or another established license)?

Regarding the old code, the modifications are rather minimal so I won't insist on that.
