### 2020-05-13 #23 @sabinomaggi
You are absolutely right, I forgot the license stuff, sorry. Now I have added a GPL3 license and changed the README file accordingly. While I was at it, I also added to the repository a directory containing the original Fortran and VB code.
Following the last requirement of the Guidelines, I have just submitted my repository to Software Heritage, and I am waiting for them to process the request.
