### 2020-05-14 #23 @pdebuyl
Thank you for the license update. In the vbprog directory, I can only read stepampl.bas which contains variable declarations. procedur.bas is a binary file that I cannot read, where should I look for the vb code for the analysis?
