### 2020-05-14 #23 @sabinomaggi
Yes you are right, all files of a Visual Basic project were binary files, you can access them only within the Visual Basic IDE.

In particular, the *.frm files  contain the layout of the graphic interface (for instance this is loadfile.frm),

![loadfile frm](https://user-images.githubusercontent.com/582340/81935561-1c240100-95f1-11ea-9030-af40c8a12fd5.png)

while the *bas files contain the actual code, but in the integrated editor you could access only one function (better procedure in basic-speak) at a time (below, you can see  the code of the FindParam function, within the file procedure.bas).

![procedur bas](https://user-images.githubusercontent.com/582340/81935845-8ccb1d80-95f1-11ea-8145-5e07fc279487.png)

Lastly, the file stepampl.mak _glued_ everything together .

![stepample mak](https://user-images.githubusercontent.com/582340/81936417-8ab58e80-95f2-11ea-9b29-8e7be7230294.png)

I have just checked and it seems there is no possibility to export the project in text form, so at the end the Visual Basic code is pretty useless outside its own environment. Another proof that VB was very immature, even by '90s standards.
