### 2020-06-03 #23 @pdebuyl
Sorry for the delay. I was also wondering the status of the resulting code, as the "updated" code (so, not the 10+ years old code) is supposed to be executable by readers of the ReScience paper. The simplest scenario is one where all code can be executed from source. In case there is a way out, I am reaching out to the editors in chief.

@ReScience/editors : Part of the reproduction involve a binary visual basic program. How does this impact the review process?
