### 2020-06-03 #23 @khinsen
@pdebuyl Badly, of course! However, if I am understanding the situation correctly, it looks like there is nothing anyone can do to fix this. So the only possible action is documenting the problem in the paper. In a sense, it's not very different from lost code.

The remaining philosophical question is if we should classify this as a reproduction failure. Not sure about that. The success/failure dichotomy isn't really appropriate.
