### 2020-06-03 #23 @sabinomaggi
@pdebuyl and @khinsen I understand your perplexities related to the Visual Basic program (the reproduction of the fortran part of the project should have no problem). The reproduction of the Visual Basic program could be easily done by building a virtual machine with MS-DOS, Windows 3.11 and Visual Basic 1.0, as explained briefly in the paper. I have not included a prebuilt VM on GitHub because I expect it would breach some license agreement, even after so many years. Clearly, I could easily share my VM upon request, but I guess it would violate the rules of the challenge.

On the other hand, the role of the Visual Basic application is only to summarize the results of the Fortran numerical simulations, calculating the size of each rf-induced step as a function of the amplitude of the microwave signal. There is nothing really interesting in that code, it only parses each data file and counts the number of occurrences of the same voltage (within a predefined range) for different values of the dc current. When writing the paper I considered all this stuff quite uninteresting, but if you wish I can try to explain it with same detail in the article and maybe add the most important parts of the code within the text.

The last possibility would be to copy the source code of each function from the VB program into a text file to be added to the repository. It is an hard work and it cannot be used anyway (Visual Basic 1.0 does not allow to import/export a source file and in any case the GUI part of the program would be left out), so probably there is not real point in doing that.
