### 2020-06-03 #23 @pdebuyl
@sabinomaggi and @khinsen thank you for the replies.

@sabinomaggi I think that the optimal solution, as Konrad mentions, is to document properly this limitation in the reproduction. It may be the case already, I'll get back to you tomorrow after re-reading the paper.
