### 2020-06-03 #23 @khinsen
Sounds good. Remember that the ultimate point of the challenge is to find out what worked and what didn't, and why. So no matter how small the role of VB in this project is, the fact that it causes trouble with open reviewing is a conclusion that matters.
