### 2020-06-04 #23 @pdebuyl
@sabinomaggi the current paper does not state clearly the the visual basic file is "imprisoned" in its binary form and unavailable in source form.

I propose the following:
1. The paper is considered a partial replication.
2. You state that the Visual Basic code is *only* available in the binary format. You already mention it ("or by opening the Visual Basic project" is not possible for readers of ReScience in the way that we expect it: open-source code that runs on current computers).
3. You include the screenshots as above in the paper (or select 1 or 2 that make the most sense).

When starting the 10-year challenge, it was hard to predict the type of results and your paper is a good illustration of the surprises that occur when resurrecting old programs.
