### 2020-06-04 #23 @sabinomaggi
No problem for 2 and 3, I will add a a note in the paper following your suggestion and include the screenshots in the Supplementary Material.
I only don't know how to mark the paper as a partial replication (or if I need to do that anyway if I comply with points 2 and 3).
