### 2020-06-11 #23 @sabinomaggi
@pdebuyl and @khinsen I have just pushed an updated version of the paper with a possible solution to the Visual Basic problem, described in a new section, Code availability. It is not perfect but is is better than having no Visual Basic code. There are also some minor changes in section 5, Visual Basic code, that are marked in color in the text, and a new Figure in the Supplementary Material showing one of the screenshots above.

I have also updated the GitHub repo of the paper, [https://github.com/sabinomaggi/ten-years-challenge-pulsed-drive](https://github.com/sabinomaggi/ten-years-challenge-pulsed-drive), adding the Basic source code and the forms that compose the GUI of the program files, that I managed to extract (somehow) from the original project.

I am not sure if this complies in full with your suggestions, but it surely improves the reproduction (and the paper).
