### 2020-06-15 #23 @rougier
Thanks for the modification and for point 1, I think what you added in the manuscript could be enough.  By the way, the type of paper is reproduction (not replication) such that the title should start with [Rp].
