### 2020-06-17 #23 @pdebuyl
@sabinomaggi sorry for the delay. It is very unusual for this type of review that I cannot execute the full reproduction. Thank you for the modifications, I like that the archive is now, in principle, complete!

@rougier I approve the paper modulo the last change that you requested about [Rp] vs [Re].
