### 2020-06-17 #23 @sabinomaggi
@pdebuyl If you wish I can share with you my DOS/Windows 3.11/Visual Basic virtual machine, maybe after converting it to the Virtual Box format (I am quite sure it is possible). Sharing privately a VM should not be too bad, what I cannot do is to make it available to everyone on a public site.
