### 2020-06-18 #23 @pdebuyl
@sabinomaggi thank you for the proposition. As it will not be ultimately part of the paper, I prefer not to go that route.
