### 2020-06-19 #23 @rougier
Based on @pdebuyl last review I think your paper is now officially accepted ! Congratulations.
For the actual publicaiton, I will need to modify your metadata to adde relevant information and I would need a softxware heritage id. Could you make one (very easy procedure, just go to the software heritage page and "save" your github repository.
