### 2020-06-19 #23 @sabinomaggi
Many thanks to @pdebuyl and @rougier for taking time to review the paper and provide great feedback.
I have just pushed an updated version of the paper with  the requested [Rp] marker. The LaTeX code has been cleaned up and I have also added the software heritage ID to the metadata (but I am not 100% sure if it is in the right format).
