### 2020-06-23 #23 @rougier
I've made a PR on your repor and  added publication information (but the DOI). Can you check everything is ok?
