### 2020-06-23 #23 @sabinomaggi
Ok, checked and merged. Should I ask for a DOI on Zenodo?
