### 2020-06-24 #23 @pdebuyl
@sabinomaggi the editor will take care of the DOI then make another PR with the updated metadata and pdf.
