### 2020-06-29 #23 @rougier
I've just added the DOI and made a PR on your repo. It's alo published on the ReScienc website with DOI: 10.5281/zenodo.3922195.

You're also on the frontpage ! You can tweet about the publication and cc @ReScienceEds such that we can retweet.

Congratulations again !
