### 2020-08-19 #23 @khinsen
@rougier Since this paper is now published, can we close this issue?
